natlab: Setup ephemeral ports only when needed
